fix : int -> long으로 수정

+++ src/main/java/gift/Repository/UserRowMapper.java
@@ -10,7 +10,7 @@ public class UserRowMapper implements RowMapper<User> {
 
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-        int id = rs.getInt("id");
+        long id = rs.getInt("id");
         String email = rs.getString("email");
         String name = rs.getString("name");
         String password = rs.getString("password");
